+++ sugar/lib/src/collection/aggregate/aggregate_iterables.dart
@@ -5,7 +5,6 @@ import 'package:sugar/core.dart';
 
 /// Provides aggregate functions, such as [average] and [sum], for [Iterable]s.
 extension AggregateIterable<E> on Iterable<E> {
-
   /// Computes the average of all values. Values are created from this iterable's elements using [select].
   ///
   /// Returns [double.nan] if empty, or if [select] returns [double.nan].
@@ -18,7 +17,8 @@ extension AggregateIterable<E> on Iterable<E> {
   ///
   /// ## Implementation details
   /// Values are computed each time rather than cached as computation is assumed to be cheap.
-  @useResult double average(Select<E, num> select) => sum(select) / length;
+  @useResult
+  double average(Select<E, num> select) => sum(select) / length;
 
   /// Computes the sum of all values. Values are created from this iterable's elements using [select].
   ///
@@ -32,7 +32,8 @@ extension AggregateIterable<E> on Iterable<E> {
   ///
   /// ## Implementation details
   /// Computing values is assumed to be cheap. Hence, values are recomputed each time rather than cached.
-  @useResult R sum<R extends num>(Select<E, num> select, {R? initial}) {
+  @useResult
+  R sum<R extends num>(Select<E, num> select, {R? initial}) {
     var sum = initial ?? 0;
     for (final element in this) {
       sum += select(element);
@@ -41,21 +42,19 @@ extension AggregateIterable<E> on Iterable<E> {
     // Dart is fucking stupid for not allowing implicit conversions between integers and doubles.
     return (R == int ? sum.toInt() : sum.toDouble()) as R;
   }
-
 }
-
 
 /// Provides ordering functions for [Iterable]s of [Comparable]s.
 ///
 /// See [OrderableIterable] for ordering elements that are not [Comparable].
 extension AggregateComparableIterable<E extends Comparable<Object>> on Iterable<E> {
-
   /// The minimum element, or `null` if empty.
   ///
   /// ```dart
   /// ['b', 'a', 'c'].min; // 'a'
   /// ```
-  @useResult E? get min {
+  @useResult
+  E? get min {
     final iterator = this.iterator;
     if (!iterator.moveNext()) {
       return null;
@@ -77,7 +76,8 @@ extension AggregateComparableIterable<E extends Comparable<Object>> on Iterable<
   /// ```dart
   /// ['a', 'c', 'b'].max; // 'c'
   /// ```
-  @useResult E? get max {
+  @useResult
+  E? get max {
     final iterator = this.iterator;
     if (!iterator.moveNext()) {
       return null;
@@ -93,7 +93,6 @@ extension AggregateComparableIterable<E extends Comparable<Object>> on Iterable<
 
     return max;
   }
-
 }
 
 /// Provides aggregate functions, such as [average] and [sum], for [Iterable]s of [num]s.
@@ -101,20 +100,21 @@ extension AggregateComparableIterable<E extends Comparable<Object>> on Iterable<
 /// This is a specialized implementation of [AggregateIterable] and [AggregateComparableIterable] that handles
 /// [double.nan] elements properly.
 extension AggregateNumberIterable<E extends num> on Iterable<E> {
-
   /// The average of all elements in this iterable, or [double.nan] if empty.
   ///
   /// ```dart
   /// [1, 2, 3].average; // 2.0
   /// ```
-  @useResult double get average => sum / length;
+  @useResult
+  double get average => sum / length;
 
   /// The sum of all elements in this iterable. Returns [double.nan] if it is present.
   ///
   /// ```dart
   /// [1, 2, 3].sum; // 6
   /// ```
-  @useResult E get sum {
+  @useResult
+  E get sum {
     num sum = 0;
     for (final element in this) {
       sum += element;
@@ -124,7 +124,6 @@ extension AggregateNumberIterable<E extends num> on Iterable<E> {
     return (E == int ? sum.toInt() : sum.toDouble()) as E;
   }
 
-
   /// The minimum element, or `null` if empty. Returns [double.nan] if it is present.
   ///
   /// ```dart
@@ -132,7 +131,8 @@ extension AggregateNumberIterable<E extends num> on Iterable<E> {
   ///
   /// [1, 2, double.nan]; // double.nan
   /// ```
-  @useResult E? get min {
+  @useResult
+  E? get min {
     final iterator = this.iterator;
     if (!iterator.moveNext()) {
       return null;
@@ -164,7 +164,8 @@ extension AggregateNumberIterable<E extends num> on Iterable<E> {
   ///
   /// [1, 2, double.nan].max; // double.nan
   /// ```
-  @useResult E? get max {
+  @useResult
+  E? get max {
     final iterator = this.iterator;
     if (!iterator.moveNext()) {
       return null;
@@ -188,5 +189,4 @@ extension AggregateNumberIterable<E extends num> on Iterable<E> {
 
     return max;
   }
-
 }