+++ sugar/lib/src/collection/maps.dart
@@ -2,7 +2,6 @@ import 'package:meta/meta.dart';
 
 /// Provides functions for working with [Map]s.
 extension Maps<K, V> on Map<K, V> {
-
   /// Merges [a] and [b], using [resolve] to resolve conflicting entries.
   ///
   /// See [putAll] for merging maps in-place.
@@ -15,7 +14,8 @@ extension Maps<K, V> on Map<K, V> {
   ///
   /// print(merged); // {'a': 1, 'b': 2}
   /// ```
-  @useResult static Map<K, V> merge<K, V>(Map<K, V> a, Map<K, V> b, {required V Function(K key, V a, V b) resolve}) {
+  @useResult
+  static Map<K, V> merge<K, V>(Map<K, V> a, Map<K, V> b, {required V Function(K key, V a, V b) resolve}) {
     final result = Map.of(a);
     for (final MapEntry(:key, :value) in b.entries) {
       final existing = a[key];
@@ -24,7 +24,6 @@ extension Maps<K, V> on Map<K, V> {
 
     return result;
   }
-
 
   /// Adds all entries in [other] to this map, using [resolve] to resolve conflicting entries.
   ///
@@ -53,9 +52,11 @@ extension Maps<K, V> on Map<K, V> {
   /// ```
   void retainWhere(bool Function(K key, V value) predicate) {
     // We remove all elements at the end to ensure atomicity.
-    [ for (final MapEntry(:key, :value) in entries) if (!predicate(key, value)) key ].forEach(remove);
+    [
+      for (final MapEntry(:key, :value) in entries)
+        if (!predicate(key, value)) key
+    ].forEach(remove);
   }
-
 
   /// Inverts this map with keys becoming values and vice versa.
   ///
@@ -63,7 +64,8 @@ extension Maps<K, V> on Map<K, V> {
   /// final foo = {'a': 1, 'b': 2, 'c': 2};
   /// foo.inverse(); //  {1: ['a'], 2: ['b', 'c']}
   /// ```
-  @useResult Map<V, List<K>> inverse() {
+  @useResult
+  Map<V, List<K>> inverse() {
     final result = <V, List<K>>{};
     for (final MapEntry(:key, :value) in entries) {
       (result[value] ??= []).add(key);
@@ -78,11 +80,11 @@ extension Maps<K, V> on Map<K, V> {
   /// final foo = {'a': 1, 'b': 2, 'c': 3};
   /// final bar = foo.where((k, v) => k == 'a' || v == 2); // {'a': 1, 'b': 2}
   /// ```
-  @useResult Map<K, V> where(bool Function(K key, V value) predicate) => {
-    for (final MapEntry(:key, :value) in entries)
-      if (predicate(key, value))
-        key: value,
-  };
+  @useResult
+  Map<K, V> where(bool Function(K key, V value) predicate) => {
+        for (final MapEntry(:key, :value) in entries)
+          if (predicate(key, value)) key: value,
+      };
 
   /// Transforms this map's keys using [convert].
   ///
@@ -97,10 +99,10 @@ extension Maps<K, V> on Map<K, V> {
   /// ```
   ///
   /// See [revalue] for converting values, and [map] for converting entries.
-  @useResult Map<K1, V> rekey<K1>(K1 Function(K key) convert) => {
-    for (final MapEntry(:key, :value) in entries)
-      convert(key): value,
-  };
+  @useResult
+  Map<K1, V> rekey<K1>(K1 Function(K key) convert) => {
+        for (final MapEntry(:key, :value) in entries) convert(key): value,
+      };
 
   /// Transforms this map's values using [convert].
   ///
@@ -110,16 +112,14 @@ extension Maps<K, V> on Map<K, V> {
   /// ```
   ///
   /// See [rekey] for converting values, and [map] for converting entries.
-  @useResult Map<K, V1> revalue<V1>(V1 Function(V value) convert) => {
-    for (final MapEntry(:key, :value) in entries)
-      key: convert(value),
-  };
-
+  @useResult
+  Map<K, V1> revalue<V1>(V1 Function(V value) convert) => {
+        for (final MapEntry(:key, :value) in entries) key: convert(value),
+      };
 }
 
 /// Provides functions for working with [Map]s of null-nullable entries.
 extension NonNullableMap<K extends Object, V extends Object> on Map<K, V> {
-
   /// Associates [key] with [value] and returns true if neither is null.
   ///
   /// Existing entries that share the same key are replaced.
@@ -141,5 +141,4 @@ extension NonNullableMap<K extends Object, V extends Object> on Map<K, V> {
 
     return result;
   }
-
 }