+++ sugar/lib/src/collection/move.dart
@@ -6,7 +6,6 @@ import 'package:sugar/core.dart';
 ///
 /// See [ListMove] for more information.
 extension MovableList<E> on List<E> {
-
   /// Returns a [ListMove] used to move elements that satisfy [where] from this list to other collections.
   ///
   /// ```dart
@@ -16,8 +15,9 @@ extension MovableList<E> on List<E> {
   /// print(foo); // [1, 3, 5]
   /// print(bar); // [2, 4]
   /// ```
-  @lazy @useResult ListMove<E> move({required Predicate<E> where}) => ListMove._(this, where);
-
+  @lazy
+  @useResult
+  ListMove<E> move({required Predicate<E> where}) => ListMove._(this, where);
 }
 
 /// A namespace for functions that move a [List]'s elements to other collections.
@@ -34,7 +34,6 @@ extension MovableList<E> on List<E> {
 ///
 /// See [SetMove] for moving a [Set]'s elements.
 class ListMove<E> {
-
   final List<E> _list;
   final Predicate<E> _predicate;
 
@@ -56,7 +55,8 @@ class ListMove<E> {
   /// print(bar); // [2, 4]
   /// ```
   @Possible({ConcurrentModificationError})
-  @useResult List<E> toList() {
+  @useResult
+  List<E> toList() {
     final moved = <E>[];
     collect(moved.add);
     return moved;
@@ -77,7 +77,8 @@ class ListMove<E> {
   /// print(bar); // {2, 4}
   /// ```
   @Possible({ConcurrentModificationError})
-  @useResult Set<E> toSet() {
+  @useResult
+  Set<E> toSet() {
     final moved = <E>{};
     collect(moved.add);
     return moved;
@@ -108,7 +109,6 @@ class ListMove<E> {
     for (final element in _list) {
       if (_predicate(element)) {
         consume(element);
-
       } else {
         retained.add(element);
       }
@@ -119,29 +119,30 @@ class ListMove<E> {
     }
 
     if (retained.length != _list.length) {
-      _list..setRange(0, retained.length, retained)..length = retained.length;
+      _list
+        ..setRange(0, retained.length, retained)
+        ..length = retained.length;
     }
   }
-
 }
 
 /// Provides functions for moving a set's elements to other collections.
 ///
 /// See [SetMove] for more information.
 extension MovableSet<E> on Set<E> {
-
   /// Returns a [SetMove] used to move elements that satisfy [where] from this [Set] to other collections.
   ///
   /// ## Example
   /// ```dart
   /// final foo = {1, 2, 3, 4, 5};
   /// final bar = foo.move(where: (e) => e.isEven).toSet();
-  /// 
+  ///
   /// print(foo); // {1, 3, 5}
   /// print(bar); // {2, 4}
   /// ```
-  @lazy @useResult SetMove<E> move({required Predicate<E> where}) => SetMove._(this, where);
-
+  @lazy
+  @useResult
+  SetMove<E> move({required Predicate<E> where}) => SetMove._(this, where);
 }
 
 /// A namespace for functions that move a [Set]'s elements to other collections.
@@ -158,7 +159,6 @@ extension MovableSet<E> on Set<E> {
 ///
 /// See [ListMove] for moving a [List]'s elements.
 class SetMove<E> {
-
   final Set<E> _set;
   final Predicate<E> _predicate;
 
@@ -178,7 +178,8 @@ class SetMove<E> {
   /// print(bar); // {2, 4}
   /// ```
   @Possible({ConcurrentModificationError})
-  @useResult Set<E> toSet() {
+  @useResult
+  Set<E> toSet() {
     final moved = <E>{};
     collect(moved.add);
     return moved;
@@ -211,5 +212,4 @@ class SetMove<E> {
 
     _set.removeAll(removed);
   }
-
 }