+++ sugar/lib/src/collection/splay_tree_maps.dart
@@ -2,7 +2,6 @@ import 'dart:collection';
 
 /// Provides functions for working [SplayTreeMap]s.
 extension SplayTreeMaps<K, V> on SplayTreeMap<K, V> {
-
   /// Returns the value associated with the first key after [key], or null if there are no keys after [key].
   ///
   /// ```dart
@@ -12,9 +11,9 @@ extension SplayTreeMaps<K, V> on SplayTreeMap<K, V> {
   /// print(map.firstValueAfter(3)); // null
   /// ```
   V? firstValueAfter(K key) => switch (firstKeyAfter(key)) {
-    null => null,
-    final key => this[key],
-  };
+        null => null,
+        final key => this[key],
+      };
 
   /// Returns the value associated with the last key before [key], or null if there are no keys before [key].
   ///
@@ -25,8 +24,7 @@ extension SplayTreeMaps<K, V> on SplayTreeMap<K, V> {
   /// print(map.lastValueBefore(1)); // null
   /// ```
   V? lastValueBefore(K key) => switch (lastKeyBefore(key)) {
-    null => null,
-    final key => this[key],
-  };
-
+        null => null,
+        final key => this[key],
+      };
 }